+++ examples/oink-oink/src/components/Game/Guessing.tsx
@@ -1,0 +1,3 @@
+export function Guessing() {
+  return <div>guessing</div>
+}
