+++ examples/oink-oink/src/components/Game/Results.tsx
@@ -1,0 +1,3 @@
+export function Results() {
+  return <div>results</div>
+}
